Fixed DOC-328:
Moved configService to separate context
Moved DB setup to backend-config.xml to be able to create genesis schema before DB configuration read
